+++ src/main/java/cz/jpalcut/pia/dao/UserDAO.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UserDAO extends JpaRepository<User, Long> {
+public interface UserDAO extends JpaRepository<User, Integer> {
 
     User findUserByLoginId(String loginId);
